Refactor hill into separate hill service
