+++ src/resolvers/WarriorResolver.ts
@@ -1,7 +1,7 @@
 import { Resolver, Mutation, ObjectType, Query, Args, Field, ArgsType, InputType } from 'type-graphql'
 import Warrior from '@/schema/Warrior'
 import Repository from '@/database/Repository'
-import MutationResult from '@/schema/MutationResult'
+import MutationResult from '@/resolvers/MutationResult'
 import { WARRIOR_COLLECTION } from '@/constants'
 import WarriorService, { IWarriorService } from '@/services/WarriorService'
