+++ src/schema/Instruction.ts
@@ -1,6 +1,6 @@
 import { ObjectType, Field } from 'type-graphql'
 import { IInstruction, OpcodeType, ModifierType } from 'corewar'
-import Operand from '@/hills/Operand'
+import Operand from '@/schema/Operand'
 
 @ObjectType()
 export default class Instruction implements IInstruction {
