+++ src/schema/Rules.ts
@@ -1,6 +1,6 @@
 import { IRules } from 'corewar'
 import { ObjectType, Field } from 'type-graphql'
-import Options from '@/hills/Options'
+import Options from '@/schema/Options'
 
 @ObjectType()
 export default class Rules implements IRules {
